# src/Contentment.Api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contentment API",
        "description": "A simple API for managing content",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/info": {
            "get": {
                "description": "Information for the current status of the API",
                "responses": {
                    "200": {
                        "description": "API Information",
                        "content": {
                            "application/vnd+contentment+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiInfo"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/content": {
            "post": {
                "description": "Create new content",
                "requestBody": {
                    "description": "Content to add to the system",
                    "required": true,
                    "content": {
                        "application/vnd+contentment+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContentCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "New content created",
                        "headers": {
                            "Location": {
                                "description": "The URL location of the newly created content",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/vnd+contentment+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Content"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/vnd+contentment-exception+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/content/{contentId}": {
            "get": {
                "description": "Get a specific piece of content",
                "parameters": [{
                    "name": "contentId",
                    "in": "path",
                    "description": "Unique content Id",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Content representation",
                        "content": {
                            "application/vnd+contentment+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Content"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiInfo": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "string",
                        "pattern": "^/d+\\./d+\\./d+$"
                    }
                }
            },
            "ContentCreate": {
                "type": "object",
                "required": [
                    "title",
                    "body"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "The content title",
                        "maxLength": 100
                    },
                    "body": {
                        "type": "string",
                        "description": "The content body"
                    }
                }
            },
            "Content": {
                "allOf": [{
                        "type": "object",
                        "required": [
                            "id",
                            "createdDateTime"
                        ],
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "The unique identifier for the content"
                            },
                            "createdDateTime": {
                                "type": "string",
                                "format": "date",
                                "description": "The date and time the content was created"
                            }
                        }
                    },
                    {
                        "$ref": "#/components/schemas/ContentCreate"
                    }
                ]
            },
            "ClientError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "detailed": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}